==> analysis/museumstation/7_main_analysis/step1_preprocess_and_merge_datasets.Rmd <==
---
title: "Drawing analyses: preprocess and merge datasets"
author: "Bria Long"
date: "7/2022"
output: html_document
---

# Libraries
```{r setup, include=FALSE}
library(tidyverse)
library(here)
library(assertthat)
library(ggthemes)
library(egg)
library(viridis)
library(lme4)
library(lmerTest)
library(MuMIn)
library(langcog)
```

## Load, describe, and join the datasets
### Filtered dataset metadata
```{r}
all_meta <- read_csv(file = here::here('data/drawings/stringent_cleaned_dataset_meta/all_object_metadata_cleaned.csv')) %>%
  as_tibble() %>%
  mutate(category = str_split_fixed(category,' ',2)[,2]) %>%
  mutate(category = str_replace(category,' ','.'))  # ice cream
```

### Basic descriptives of filtered set
```{r}
num_subs = length(unique(all_meta$session_id))
```

The final, filtered dataset of N=`r length(unique(all_meta$filename))`  drawings from `r length(unique(all_meta$category))` categories from `r num_subs` children who were on average `r mean(all_meta$age_numeric)`years of age (range 3-10 years). 

### Load category frequency ratings
```{r}
frequency = read_csv(file = here::here('data/surveys/drawing_experience/preprocessed/Category_frequency_survey.csv')) 
```

To assess this, `r length(unique(frequency$subject_id))` parents of children aged 3-10 years filled out a survey asking about the frequency with with their children drew the categories in the dataset.

```{r}
count_by_age <- frequency %>%
  group_by(childs_age) %>%
  dplyr::summarize(num_surveys = length(unique(subject_id)))
```

```{r}
freq_by_category <- frequency %>%
  # filter(childs_age > 2) %>%
  mutate(category = str_split_fixed(category,' ',2)[,2]) %>%
  mutate(category = str_replace(category,' ','.')) %>%
  filter(category %in% all_meta$category) %>%
  group_by(category) %>%
  summarize(drawing_frequency = mean(often_drawn_rating)) %>%
  mutate(above_median_freq = drawing_frequency > median(drawing_frequency)) 

write_csv(freq_by_category, here::here('data/surveys/drawing_experience/preprocessed/freq_by_category.csv'))


```

### Load tracing data
```{r}
all_tracings <- read_csv(here('data/tracing/rated_all_museumstation_filtered.csv'))%>%
  select(-...1, -X)

## Make averages for joining
by_subject_tracing_avg <- all_tracings %>%
  distinct(session_id, category, age, rating) %>% 
  group_by(session_id) %>%
  summarize(avg_tracing_rating = mean(rating))
```

### Load category animacy/size (hardcoding)
```{r}
animacy_csv <- read_csv(here::here('data/drawings/category_metadata/animacy.csv')) %>%
  as_tibble() %>%
  mutate(animacy_size = case_when(animacy == '0' & size=='0' ~ 2, # small objects
                                  animacy == '0' & size=='1' ~ 1, # big objects
                                  animacy == '1' & size=='0' ~ 4, # small animals
                                  animacy == '1' & size=='1' ~ 3)) # big animals

```

### Load CLIP classifications
```{r}
load(file=here::here('data/clip_outputs/fullset/clip_by_image.RData'))
load(file=here::here('data/clip_outputs/fullset/clip_long.RData'))

```


### Load VGG classification data
```{r}
num_batches=232
# reg_string = 'C_0.1_T_0.1'
reg_string = 'P1'

layers = c('P1','P2','P3','P4','P5','FC6','FC7')
for (this_layer in layers){
  
temp_class_data <- read.csv(here::here('data','compiled_classifications',paste0(this_layer,'batchtotal_', as.character(num_batches),'.csv'))) %>%
  mutate(session_id = paste('cdm_',session_id,sep="")) %>%
  mutate(age_numeric = age) %>%
  mutate(age = paste('age',age,sep="")) %>%
  mutate(age = as.factor(age)) %>%
  mutate(category = target_label) %>%
  mutate(image_name = paste(target_label,'_sketch_', age,'_', session_id,'.png',sep="")) %>%
  select(-X)  %>%
  mutate(category = str_replace(category,' ','.')) %>% # ice cream = ice.cream 
  mutate(layer = this_layer) 

  if (this_layer=='P1'){
    all_layer_class = temp_class_data
  }
  else {
  all_layer_class <- all_layer_class %>%
    full_join(temp_class_data)
  }

}
```


```{r}
## Order these layers reasonably.
layers_ordered = c('P1','P2','P3','P4','P5','FC6','FC7')

all_layer_class <- all_layer_class %>%
 mutate(layer = factor(layer, levels = layers_ordered)) %>%
 select(-X.1, -index, -batch, -batch_str) %>%
 rename(guessed_category = target_label)


```


#### For each layer
Save vgg layers outputs for extra analyses
```{r}
save(all_layer_class, file=here::here('data/compiled_classifications/vgg_outputs.RData'))
```


```{R}
# classification_data_fc7 <- read.csv(here::here('data','compiled_classifications/',paste0('C_0.1_T_0.1', 'batchtotal_',as.character(num_batches),'.csv'))) %>%
#   mutate(session_id = paste('cdm_',session_id,sep="")) %>%
#   mutate(age_numeric = age) %>%
#   mutate(age = paste('age',age,sep="")) %>%
#   mutate(age = as.factor(age)) %>%
#   mutate(category = target_label) %>%
#   mutate(image_name = paste(target_label,'_sketch_', age,'_', session_id,'.png',sep="")) %>%
#   select(-X)  %>%
#   mutate(category = str_replace(category,' ','.'))  # ice cream = ice.cream

```

```{r}
# classification_data_fc6 <- read.csv(here::here('data','compiled_classifications/',paste0('FC6','batchtotal_',as.character(num_batches),'.csv'))) %>%
#   mutate(session_id = paste('cdm_',session_id,sep="")) %>%
#   mutate(age_numeric = age) %>%
#   mutate(age = paste('age',age,sep="")) %>%
#   mutate(age = as.factor(age)) %>%
#   mutate(category = target_label) %>%
#   mutate(image_name = paste(target_label,'_sketch_', age,'_', session_id,'.png',sep="")) %>%
#   select(-X)  %>%
#   mutate(category = str_replace(category,' ','.'))  # ice cream = ice.cream

```


### Load part emphasis
```{r}
# all annotations before filtering
load(file = here::here('data/part_annotations_processed/merged_annotations.RData'))
```

```{r}
# preprocesssed emphasis data
load(file = here::here('data/part_annotations_processed/part_emphasis.RData'))
```

```{r}
# example drawing in presentation, or any other we want to examine
# example <- part_emphasis_all %>% filter(filename == 'bird_sketch_age8_cdm_run_v31530484377182.png')
# 
# what_annotations <- d_agree_labels %>% filter(filename == 'bird_sketch_age8_cdm_run_v31530484377182.png')
```

## Join meta data, drawing frequnecy, and classification data
```{r}
d <- all_layer_class %>% 
  filter(layer=='FC6') %>%
  gather(key = 'class', value = 'prob', contains('prob')) %>%
  mutate(class = str_split_fixed(class, '_prob',2)[,1]) %>%
  group_by(image_name, age, category, correct_or_not, session_id, age_numeric) %>%
  summarize(denom = sum(prob), target_label_prob = prob[class==category], log_odds = log(target_label_prob / (denom - target_label_prob))) %>%
  rename(filename = image_name) %>%
  left_join(all_meta, by=c("filename", "category", "age_numeric","session_id")) %>%
  mutate(draw_duration = draw_duration_old) %>% # using end of last stroke - end of first stroke (is 0 for one stroke drawings)
  mutate(run = substr(session_id,0,10)) %>% 
  left_join(freq_by_category) %>%
  mutate(correct_or_not = as.numeric(as.logical(correct_or_not)))

```


## Join tracing ratings
```{r}
d <- d %>%
  left_join(by_subject_tracing_avg)
```


### Asserts to check all the joins
```{r}
# weird things were happening with category matching, check
assert_that(length(unique(d$filename)) == length(unique(all_layer_class$image_name)))

# every drawing should have all of these, regardless
assert_that(sum(is.na(d$age_numeric))==0)
assert_that(sum(is.na(d$category))==0)
assert_that(sum(is.na(d$correct_or_not))==0)

missing_meta <- d %>%
  filter(is.na(num_strokes))

assert_that(length(missing_meta$filename)==0)
```

```{r}
## merged vgg classifications with meta
save(d, file=here::here('data/preprocessed_data/merged_vgg_class_and_meta.RData'))

##
write_csv(d, file=here::here('data/preprocessed_data/merged_vgg_class_and_meta.csv'))

```


# Merge CLIP classificaitons wtih othe rdata

```{R}
clip_by_image <- clip_long %>%
  mutate(category = str_replace(category,' ','.')) %>%
  mutate(guessed_category = str_replace(guessed_category,' ','.'))  %>%
  mutate(category = as.factor(category), guessed_category = as.factor(guessed_category)) %>%
  group_by(kid_age, session_id, category) %>%
  summarize(max_prob = max(prob), clip_category = guessed_category[prob == max_prob]) 
# %>%
  # mutate(correct_or_not = (category == clip_category))
```

```{r}
# There are 695 drawings for which clip isn't sure between 2 options 
what <- clip_by_image %>%
  group_by(kid_age, session_id, category) %>%
  mutate(num_matches = length(clip_category)) %>%
  filter(num_matches > 1) 
```

```{r}
clip_by_image_better <- clip_long %>%
  rename(age_numeric = kid_age) %>%
  mutate(category = str_replace(category,' ','.')) %>%
  mutate(guessed_category = str_replace(guessed_category,' ','.'))  %>%
  group_by(session_id, category, age_numeric) %>%
  summarize(target_label_prob = prob[category==guessed_category], denom = sum(prob),  log_odds = log(target_label_prob / (denom - target_label_prob)), max_prob = max(prob), clip_category = guessed_category[prob == max_prob])  %>%
  mutate(correct_or_not = (clip_category == category)) %>%
  group_by(age_numeric, session_id, category) %>%
  sample_n(1) 

```


```{r}
d_clip <- clip_by_image_better %>% 
  mutate(session_id = paste0("cdm_", session_id)) %>%
  group_by(age_numeric, category, correct_or_not, session_id) %>%
  left_join(all_meta, by=c("category", "age_numeric","session_id")) %>%
  mutate(draw_duration = draw_duration_old) %>% # using end of last stroke - end of first stroke (is 0 for one stroke drawings)
  mutate(run = substr(session_id,0,10)) %>% 
  left_join(freq_by_category) %>%
  mutate(correct_or_not = as.numeric(as.logical(correct_or_not))) %>%
  left_join(by_subject_tracing_avg)


```

### Asserts to check all the joins
```{r}
# weird things were happening with category matching, check
# assert_that(length(unique(d_clip$filename)) == length(unique(all_layer_class$image_name)))

# every drawing should have all of these, regardless
assert_that(sum(is.na(d_clip$age_numeric))==0)
assert_that(sum(is.na(d_clip$category))==0)
assert_that(sum(is.na(d_clip$correct_or_not))==0)

missing_meta <- d_clip %>%
  filter(is.na(num_strokes))

assert_that(length(missing_meta$filename)==0)
```

```{r}
## merged clip classifications with meta
save(d_clip, file=here::here('data/preprocessed_data/merged_clip_class_and_meta.RData'))

##
write_csv(d_clip, file=here::here('data/preprocessed_data/merged_clip_class_and_meta.csv'))
```

```{r}

```

# Main Data Structures

Main data structures for analysis:
##
d: Has VGG FC6 classifications, effort covariates, tracing accuracy

d_clip: has clip classificaitons, effort covariates, tracing accuracy

##
all_layer_class -- has all vgg layer classifications for every image (but not other covariates)

clip_long: clip long form classificatinos (each image x 48 categories)

##
d_agree_labels: part labels that were agreed upon by annotators for 2088 drawings



## Figure XX. Changes in part structure by age

```{r}
img_per_category_age <- d_agree_labels %>%
  group_by(category, age_numeric) %>%
  summarize(count_images = length(unique(filename)))  %>%
  mutate(category = factor(category, levels = c("bird", "fish", "rabbit","dog" ,"sheep","tiger" ,"camel" , "bear", "cup","lamp", "hat","bottle", "car", "boat","airplane","train")))

top_parts_emphasis <- part_emphasis_all %>%
  group_by(category, roi_labelName) %>%
  filter(!roi_labelName %in% c('body','head')) %>%
  summarize(avg_emphasis = mean(emphasis), count_drawings = n()) %>%
  group_by(category) %>%
  slice_max(n=5, order_by=count_drawings)

part_emphasis_top_4_plot <- part_emphasis_all %>%
  mutate(diag_feature_annotation = (roi_labelName == diag_feature)) %>%
  right_join(top_parts_emphasis %>% select(category, roi_labelName)) %>%
  group_by(category, age_numeric, roi_labelName, diag_feature_annotation) %>%
  dplyr::summarize(avg_emphasis = mean(emphasis), count_drawings_with_part = n()) %>%
  left_join(img_per_category_age) %>%
  mutate(prop_drawings = count_drawings_with_part / count_images)
  

```

```{r}

```


```{r}

ggplot(data = part_emphasis_top_4_plot, aes(x=age_numeric, y=prop_drawings, col=roi_labelName)) +
  geom_point(aes(size = avg_emphasis, alpha=.2)) +
  scale_size(range = c(0,4)) +
  facet_wrap(~category, nrow=1) +
  theme_few(base_size=12, base_family = "Helvetica") +
  geom_line(stat='smooth',alpha=.2, span=10, aes(group=roi_labelName)) +
  ylab('Proportion drawings with object part') +
  xlab('Age of drawer') +
  theme(legend.position = 'none', aspect.ratio = 1)  +
  ylim(0,1) +
  facet_wrap(~category, nrow=2, scales = 'free_x') +
  ggrepel::geom_label_repel(data =part_emphasis_top_4_plot %>% filter(age_numeric==8), aes(label = roi_labelName), max.overlaps=10, size=2, label.size=.1, box.padding=.1)

ggsave('part_emphasis_by_age.pdf',units = 'in', height = 4, width = 9)

```


## Appendix Figure: Changes in object part structure with 48-way classification
```{r}
part_emphasis_dnn_recognition <- part_emphasis_all %>%  
  mutate(sketch_path = as.factor(str_split_fixed(filename,'.png',2)[,1])) %>%
  left_join(d)  %>%
  filter(!is.na(correct_or_not))
```

```{r}
# A bunch of drawings were ecxluded from dnn classifications for balancing...annoying

# what <- part_emphasis_all %>%  
#   mutate(sketch_path = as.factor(str_split_fixed(filename,'.png',2)[,1])) %>%
#   left_join(d)  %>%
#   filter(is.na(correct_or_not)) %>%
#   ungroup() %>%
#   distinct(filename, category, age_numeric) %>%
#   group_by(category, age_numeric) %>%
#   summarize(num_excluded = n())

```

```{r}
part_emphasis_recognition_binned <-  part_emphasis_dnn_recognition %>%
  ungroup() %>%
  group_by(category) %>% # construct bins WITHIN categories
  mutate(model_correct_bin = ntile(log_odds,4))


```


```{r}
binned_count_recognition_by_sketch <- part_emphasis_recognition_binned %>%
  group_by(category, model_correct_bin) %>%
  summarize(count_images_bin = length(unique(filename)))

part_emphasis_recognition_binned <- part_emphasis_recognition_binned %>%
  left_join(binned_count_recognition_by_sketch)
```

```{r}
# get emphasized parts without recognition
top_parts_emphasis <- part_emphasis_all %>%
  group_by(category, roi_labelName) %>%
  # filter(!roi_labelName %in% c('body','head')) %>%
  summarize(avg_emphasis = mean(emphasis), count_drawings = n()) %>%
  group_by(category) %>%
  slice_max(n=5, order_by=count_drawings)
#
```

```{r}
part_emphasis_top_4_by_model_bin <- part_emphasis_recognition_binned %>%
  right_join(top_parts_emphasis %>% select(category, roi_labelName)) %>%
  group_by(category, model_correct_bin, roi_labelName,count_images_bin) %>%
  dplyr::summarize(avg_emphasis = mean(emphasis), count_drawings_with_part = length(unique(filename))) %>%
  mutate(prop_drawings = count_drawings_with_part / count_images_bin)
```


```{r}
# part_names = unique(part_emphasis_top_4_by_model_bin$roi_labelName)
part_emphasis_top_4_by_model_bin <- part_emphasis_top_4_by_model_bin %>%
  # mutate(label = as.factor(roi_labelName), levels =  unique(roi_labelName)) %>%
  mutate(category = factor(category, levels = c("bird", "fish", "rabbit","dog" ,"sheep","tiger" ,"camel" , "bear", "cup","lamp", "hat","bottle", "car", "boat","airplane","train")))
```

```{r}
ggplot(data = part_emphasis_top_4_by_model_bin, aes(x=model_correct_bin, y=prop_drawings, col=roi_labelName)) +
  geom_point(aes(size = avg_emphasis, alpha=.2)) +
  # geom_linerange(aes(ymin = ci_lower, ymax = ci_upper), alpha=.5) +
  facet_wrap(~category, nrow=1) +
  theme_few(base_size=8) +
  # geom_line() +
  geom_line(stat='smooth',alpha=.2, span=10) +
  # stat_smooth(geom="line", alpha=0.3, size=.5, span=10) +
  # geom_smooth(aes(group=roi_labelName), span=10, alpha=.01) +
  # scale_color_viridis(option = 'F', discrete=TRUE) +
  ylab('Proportion drawings with object part') +
  xlab('Binned classifier evidence (VGG-19 classifications)') +
  theme(legend.position = 'none', aspect.ratio = 1)  +
  ylim(0,1) +
  facet_wrap(~category, nrow=2, scales = 'free_x') +
  scale_x_continuous(
    limits=c(.5,4.5),
    breaks=c(1,4),
    labels=c(' Low ',' High ')
        ) +
  ggrepel::geom_label_repel(data =part_emphasis_top_4_by_model_bin %>% filter(model_correct_bin==1), aes(label = roi_labelName), max.overlaps=10, size=2, box.padding=.1, label.padding=.1)

ggsave('part_emphasis_by_model_bin.pdf',units = 'in', height = 3, width = 8)
```

## Figure 3: Misclassified draiwngs by VGG (+ each layer) + CLIP
### Make long form VGG classification data
```{r}
classification_data_long <- all_layer_class %>%
  mutate(correct_or_not = as.logical(correct_or_not)) %>%
  gather(key = 'class', value = 'prob', contains('prob')) %>%
  mutate(class = str_split_fixed(class, '_prob',2)[,1]) 
```

### Make long form clip for misclassified data
```{r}
clip_misclassified_long <- clip_long %>%
  right_join(clip_by_image %>% filter(correct_or_not==FALSE) %>% select(-max_prob, -clip_category))  %>%
  mutate(guessed_category = str_replace(guessed_category,' ','.')) %>% # ice cream = ice.cream %>%
  mutate(category = str_replace(category,' ','.'))
```


### Calculate avg prob by category for misclassified

```{r}
#### For VGG
confusions_by_class_by_layer <- classification_data_long %>%
  mutate(drawn_category = category) %>%
  left_join(animacy_csv) %>%
  ungroup() %>%
  filter(correct_or_not==0) %>%
  mutate(drawn_category = fct_reorder(drawn_category, animacy_size)) %>%
  mutate(class = factor(class, levels = levels(drawn_category))) %>%
  group_by(drawn_category, class, layer) %>%
  dplyr::summarize(mean_prob = mean(prob)) 
```

```{r}
#### For CLIP
clip_confusions_by_class<- clip_misclassified_long %>%
  mutate(drawn_category = as.factor(category)) %>%
  left_join(animacy_csv) %>%
  ungroup() %>%
  mutate(drawn_category = fct_reorder(drawn_category, animacy_size)) %>%
  mutate(guessed_category = factor(guessed_category, levels = levels(drawn_category))) %>%
  group_by(drawn_category, guessed_category,animacy_size) %>%
  dplyr::summarize(mean_prob = mean(prob)) 

```

### Appendix FIgure: Clip confusions by class
```{r}
ggplot(data = clip_confusions_by_class, aes(x=guessed_category, y=drawn_category,fill=mean_prob)) + 
  geom_tile() + 
  theme_few(base_size = 12) +
  theme(legend.position = 'left', axis.text.x = element_text(angle = 90,  
  size = 6, vjust=.5, hjust = 1), axis.text.y = element_text(angle = 0,
  size = 6, hjust = 1)) + 
  coord_fixed() + 
  scale_fill_viridis(option="A", name = 'Classifier probability')  +
  ylab('Drawn category')  +
  xlab('Predicted category')


ggsave('figures/appendix_clip_classifier_confusions.pdf', width = 4, height = 5, units = 'in')
```
### Appendix Figure: VGG confusions by class for each layer
```{r}
ggplot(data = confusions_by_class, aes(x=class, y=drawn_category,fill=mean_prob)) + 
  geom_tile() + 
  theme_few(base_size = 12) +
  theme(legend.position = 'left', axis.text.x = element_text(angle = 90,  
  size = 6, vjust=.5, hjust = 1), axis.text.y = element_text(angle = 0,
  size = 6, hjust = 1)) + 
  coord_fixed() + 
  scale_fill_viridis(option="A", limits=c(0,quantile(confusions_by_class$mean_prob,.998)), name = 'Classifier probability')  +
  ylab('Drawn as') +
  xlab('Confused with') +
  facet_grid(~layer)

# ggsave('figures/classifier_confusions.pdf', width = 4, height = 5, units = 'in')
```



### Calculate whether animacy/size were correct
```{r}
animacy_size_acc_by_age <- classification_data_long %>%
  mutate(drawn_category = category) %>%
  left_join(animacy_csv %>% select(animacy, size, category), by = c("drawn_category" = "category")) %>%
  ungroup() %>%
  rename(drawn_category_animacy = animacy, drawn_category_size = size) %>%
  left_join(animacy_csv %>% select(animacy, size, category),  by = c("class" = "category")) %>%
  rename(class_animacy = animacy, class_size = size) %>%
  group_by(image_name, drawn_category, age_numeric, layer) %>%
  mutate(max_prob = max(prob), top_class = prob==max_prob)  %>%
  filter(top_class==TRUE) %>%
  mutate(animacy_correct = (drawn_category_animacy == class_animacy)) %>%
  mutate(size_correct = (drawn_category_size == class_size)) 
```


```{r}

clip_animacy_size_acc_by_age <- clip_misclassified_long %>%
  mutate(drawn_category = category) %>%
  left_join(animacy_csv %>% select(animacy, size, category), by = c("drawn_category" = "category")) %>%
  ungroup() %>%
  rename(drawn_category_animacy = animacy, drawn_category_size = size) %>%
  left_join(animacy_csv %>% select(animacy, size, category),  by = c("guessed_category" = "category")) %>%
  rename(class_animacy = animacy, class_size = size) %>%
  group_by(session_id, drawn_category, kid_age) %>%
  mutate(max_prob = max(prob), top_class = prob==max_prob)  %>%
  filter(top_class==TRUE) %>%
  mutate(animacy_correct = (drawn_category_animacy == class_animacy)) %>%
  mutate(size_correct = (drawn_category_size == class_size)) 

```

### Calculate animacy correct relative to basline
```{r}
baseline_animacy <- mean(animacy_csv$animacy)
baseline_objects <- 1 - mean(animacy_csv$animacy)
```

```{r}
## for vgg
anim_cor_by_category_by_age <- animacy_size_acc_by_age %>%
  filter(correct_or_not==0) %>%
  mutate(drawn_category_animacy = as.factor(drawn_category_animacy))  %>%
  mutate(baseline_chance = case_when(drawn_category_animacy == 0 ~ baseline_objects, 
                                     drawn_category_animacy == 1 ~ baseline_animacy)) %>%
  group_by(age_numeric, drawn_category, layer) %>%
  dplyr::summarize(avg_animacy_correct = mean(animacy_correct) - baseline_chance) %>%
  distinct(age_numeric, avg_animacy_correct, drawn_category)
  
anim_cor_by_age <- anim_cor_by_category_by_age %>%
  group_by(age_numeric, layer) %>%
  multi_boot_standard(col = 'avg_animacy_correct')

```


```{r}
## for clip
clip_anim_cor_by_category_by_age <- clip_animacy_size_acc_by_age %>%
  rename(age_numeric = kid_age) %>%
  filter(correct_or_not==0) %>%
  mutate(drawn_category_animacy = as.factor(drawn_category_animacy))  %>%
  mutate(baseline_chance = case_when(drawn_category_animacy == 0 ~ baseline_objects, 
                                     drawn_category_animacy == 1 ~ baseline_animacy)) %>%
  group_by(age_numeric, drawn_category) %>%
  dplyr::summarize(avg_animacy_correct = mean(animacy_correct) - baseline_chance) %>%
  distinct(age_numeric, avg_animacy_correct, drawn_category)
  
clip_anim_cor_by_age <- clip_anim_cor_by_category_by_age %>%
  group_by(age_numeric) %>%
  multi_boot_standard(col = 'avg_animacy_correct')

```



### Figure 3B: Plot animacy correct by age
```{r}
ggplot(anim_cor_by_age, aes(x=age_numeric,y=mean, color=age_numeric)) +
  geom_pointrange(aes(ymin = ci_lower, ymax = ci_upper), alpha=.8, size=.2) + 
  geom_point(data=anim_cor_by_category_by_age, aes(x=age_numeric, y=avg_animacy_correct), alpha=.1, size=.5) +
  theme_few(base_size = 8) + 
  labs(x='Age', y='Prop. animacy correct') +
  theme(legend.position = "none") +
  geom_smooth(span=20, alpha=smooth_alpha, color='grey') +
  scale_color_viridis(option="D", discrete=FALSE) +
  geom_hline(yintercept =  0 , linetype="dashed", color="grey")  +
  facet_grid(~layer)


ggsave('figures/animacy_classification.pdf', width = 1.5, height = 1.5, units = 'in')


```

```{r}
objects_only <- animacy_csv %>%
  filter(animacy==0)

baseline_big <- mean(objects_only$size)
baseline_small <- 1 - mean(objects_only$size)

size_cor_by_category_by_age <- animacy_size_acc_by_age %>%
  filter(correct_or_not==0) %>%
  filter(drawn_category_animacy==0) %>% #only objects
  mutate(drawn_category_size = as.factor(drawn_category_size))  %>%
  mutate(baseline_chance = case_when(drawn_category_size == 0 ~ baseline_small, 
                                     drawn_category_size == 1 ~ baseline_big)) %>%
  group_by(age_numeric, drawn_category, layer) %>%
  dplyr::summarize(avg_size_correct = mean(size_correct) - baseline_chance) %>%
  distinct(age_numeric, avg_size_correct, drawn_category)

size_cor_by_age <- size_cor_by_category_by_age %>%
  group_by(age_numeric, layer) %>%
  multi_boot_standard(col = 'avg_size_correct')

```

```{r}
ggplot(clip_size_cor_by_age, aes(x=age_numeric,y=mean, color=age_numeric)) +
  geom_pointrange(aes(ymin = ci_lower, ymax = ci_upper), alpha=.8, size=.2) + 
  geom_point(data=clip_size_cor_by_category_by_age, aes(x=layer, y=avg_size_correct), alpha=.1, size=.5) +
  theme_few(base_size = 8) + 
  labs(x='Age', y='Prop. object size correct') +
  theme(legend.position = "none") +
  # geom_smooth(span=10, alpha=smooth_alpha, color='grey') +
  # scale_color_viridis(option="D", discrete=FALSE) +
  geom_hline(yintercept =  0 , linetype="dashed", color="grey")  
```

```{r}
ggplot(clip_anim_cor_by_age, aes(x=age_numeric,y=mean, color=age_numeric)) +
  geom_pointrange(aes(ymin = ci_lower, ymax = ci_upper), alpha=.8, size=.2) + 
  geom_point(data=clip_anim_cor_by_category_by_age, aes(x=age_numeric, y=avg_animacy_correct), alpha=.1, size=.5) +
  theme_few(base_size = 8) + 
  theme(aspect.ratio=1) +
  labs(x='Age', y='Clip prop animacy correct') +
  theme(legend.position = "none") +
  geom_smooth(span=10, alpha=.3, color='grey') +
  scale_color_viridis(option="D", discrete=FALSE) +
  geom_hline(yintercept =  0 , linetype="dashed", color="grey")  

```

### Figure Anim and Size correct by age by each vgg layer
```{r}
ggplot(anim_cor_by_age, aes(x=layer,y=mean, color=age_numeric)) +
  geom_pointrange(aes(ymin = ci_lower, ymax = ci_upper), alpha=.8, size=.2) + 
  geom_point(data=size_cor_by_category_by_age, aes(x=layer, y=avg_size_correct), alpha=.1, size=.5) +
  theme_few(base_size = 8) + 
  labs(x='Age', y='Prop. object size correct') +
  theme(legend.position = "none") +
  # geom_smooth(span=10, alpha=smooth_alpha, color='grey') +
  # scale_color_viridis(option="D", discrete=FALSE) +
  geom_hline(yintercept =  0 , linetype="dashed", color="grey")  +
  facet_grid(~age_numeric)

ggplot(size_cor_by_age, aes(x=layer,y=mean, color=age_numeric)) +
  geom_pointrange(aes(ymin = ci_lower, ymax = ci_upper), alpha=.8, size=.2) + 
  geom_point(data=size_cor_by_category_by_age, aes(x=layer, y=avg_size_correct), alpha=.1, size=.5) +
  theme_few(base_size = 8) + 
  labs(x='Age', y='Prop. object size correct') +
  theme(legend.position = "none") +
  # geom_smooth(span=10, alpha=smooth_alpha, color='grey') +
  # scale_color_viridis(option="D", discrete=FALSE) +
  geom_hline(yintercept =  0 , linetype="dashed", color="grey")  +
  facet_grid(~age_numeric)
```


```{r}

# ggsave('figures/size_classification.pdf', width = 1.5, height = 1.5, units = 'in')
```





# Run inferential statisitics and output
```{r}
# set seed for reproducibility
set.seed(123)
```


## Full model predicting accuracy for all drawings
```{R}
all_covariates_accuracy <- glmer(correct_or_not ~ scale(age_numeric)*scale(drawing_frequency)+scale(avg_tracing_rating) + scale(draw_duration) +
                          scale(mean_intensity) +
                          scale(num_strokes) +
                        (1|session_id) +
                        (1|category),
      data = d, family="binomial",control=glmerControl(optimizer="bobyqa",optCtrl=list(maxfun=2e4)))
```

```{R}
no_tracing <- glmer(correct_or_not ~ scale(age_numeric)*scale(drawing_frequency)+ scale(draw_duration) +
                          scale(mean_intensity) +
                          scale(num_strokes) +
                        (1|session_id) +
                        (1|category),
      data = d, family="binomial",control=glmerControl(optimizer="bobyqa",optCtrl=list(maxfun=2e4)))
```

```{r}
no_tracing_compare <- glmer(correct_or_not ~ scale(age_numeric)*scale(drawing_frequency)+ scale(draw_duration) +
                          scale(mean_intensity) +
                          scale(num_strokes) +
                        (1|session_id) +
                        (1|category),
      data = d %>% filter(!is.na(avg_tracing_rating)), family="binomial",control=glmerControl(optimizer="bobyqa",optCtrl=list(maxfun=2e4)))


tracing_compare <- glmer(correct_or_not ~ scale(age_numeric)*scale(drawing_frequency)+ scale(avg_tracing_rating) + scale(draw_duration) +
                          scale(mean_intensity) +
                          scale(num_strokes) +
                        (1|session_id) +
                        (1|category),
      data = d %>% filter(!is.na(avg_tracing_rating)), family="binomial",control=glmerControl(optimizer="bobyqa",optCtrl=list(maxfun=2e4)))
```

```{r}
anova(tracing_compare, no_tracing_compare)
```

# Check multicolinearity
```{r}
# https://easystats.github.io/blog/posts/performance_check_collinearity/#:~:text=Multicollinearity%20%E2%80%9Cis%20a%20phenomenon%20in,as%20non%2Dindependent%20covariates).
library(performance)
check_collinearity(all_covariates_accuracy)
```

### Table 1
```{R}
xtable::xtable(summary(all_covariates_accuracy)$coef, digits=3, caption = "Model coefficients of a GLMM predicting the recognizability of each drawing (i.e. binary classification scores), including randomintercepts for each category and participant.")

```

# Full model predicting log odds for correctly classified drawings
```{r}
correct_subset <- d %>% 
  filter(correct_or_not==1)

all_covariates_log_odds_corr_only <- lmer(log_odds ~ scale(age_numeric)*scale(drawing_frequency) + scale(avg_tracing_rating) + scale(draw_duration) +
                          scale(mean_intensity) +
                          scale(num_strokes) +
                        (1|session_id) +
                        (1|category),
      data = correct_subset, control=lmerControl(optimizer="bobyqa",optCtrl=list(maxfun=2e4)))

log_odds_out = summary(all_covariates_log_odds_corr_only)$coef

```

Even for drawings that were correctly classified (near 33\% of the balanced subset of drawings, $N$=`r length(unique(correct_subset$filename))`) there was still a reliable increase in the amount of diagnostic information they contained across age, as measured by the log-odds probability assigned by the logistic-regression classifier to the target category (see \textit{Methods}, \textit{B} = `r round(log_odds_out[2,1],3)` , \textit{SE} = `r round(log_odds_out[2,2],3)`, \textit{df} =  `r round(log_odds_out[2,3],3)`, \textit{t} = `r round(log_odds_out[2,4],3)`, \textit{P} < `r round(log_odds_out[2,5],3)`). Thus, this secondary analysis provides additional support for the account that the age-related improvements reflect gradual changes in children's ability to include diagnostic visual features in their drawings (see Figure \ref{fig:drawing_production}B).
